Fixed errors caused by the new Rust nightly
The most recent Rust nightly stabilized `TryFrom` and `TryInto` (which is good)
and added them to the `prelude` (also good) which broke a lot of my code because
of the implementations of `try_from` and `try_into` in `MaybeFrom` and
`MaybeInto`, which became name clashes. The easy solution was to rename those
methods, which are now called `try_convert_from` and `try_convert_into`.